Implement Automatic Backups
Images are now stored in two places - “cases” and “backups”. Those in
the former folder are live and include all changes made by officers,
those in the latter are versioned so that a new file is created every
time an officer edits and saves. Format used is “image.jpg” -> “image
(1).jpg” - > “image (2).jpg” -> etc.