Major Rewrite
Designed an object oriented system for dealing with and representing
evidence files stored on the local device. Worked on detecting images
on peripheral devices. Generally cleaned up the code and made it much
more understandable and efficient. Converted standard comments to
JavaDoc comments in order to make development easier. Working on making
the graphical user interface work with all of these changes.