Print Works (Almost)
Finally got the generated PDF to print properly. Added options allowing
the user to select print properties such as printer name, number of
copies, and page orientation. Added additional functions to
ComponentGenerator. Please make note of a bug affecting OS X users, in
which the program always prints to the default printer regardless of
what the user chooses. This is well documented in the OpenJDK bug
tracker, and is not a PEMS issue but rather a Java related issue.
Thankfully, the PPD is using Windows.